requirements.txt
aiogram==2.24
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==2.1.1
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
python-dotenv==0.21.0
pytz==2022.7.1
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
urllib3==1.26.14
vk-api==11.9.9
yarl==1.8.2
